net: misc updates

+++ src/net/common.rs
@@ -32,16 +32,16 @@ pub const MAX_TX_REQUEST: u32 = 10_000;
 pub const MAX_CLAIM_REQUEST: u32 = 1_000;
 
 /// Service constant for Network capabilities (1 << 0)
-pub const NETWORK: u64 = 1;
+pub const NETWORK: u32 = 1;
 
 ///Service constant for Bloom Filter capabilities
-pub const BLOOM: u64 = (1 << 1);
+pub const BLOOM: u32 = (1 << 1);
 
 ///Service definition for a full node - (3)
-pub const FULL_NODE: u64 = NETWORK | BLOOM;
+pub const FULL_NODE: u32 = NETWORK | BLOOM;
 
 ///Service definition required to communicate - (3)
-pub const REQUIRED_SERVICES: u64 = NETWORK | BLOOM;
+pub const REQUIRED_SERVICES: u32 = NETWORK | BLOOM;
 
 ///Service definition for rsd - (3)
-pub const LOCAL_SERVICES: u64 = NETWORK | BLOOM;
+pub const LOCAL_SERVICES: u32 = NETWORK | BLOOM;
